+++ src/main/java/app/entities/IEntity.java
@@ -1,5 +1,5 @@
 package app.entities;
 
 public interface IEntity {
-    Long getID();
+    int getID();
 }
